Add ARROW_RESULTS client capability
